+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Drawer: Added new component
 
+### Fixed
+- DropDownSelect: Fixed long strings display
+
 
 ## [0.42.0] - 2021-03-15